Migrate Mutation.createProject, abstract validation rules
